Removed the frame options
When you specify frame options like 'color' the window doesn't inherit the OS' theme in regards to window controls, colours, etc. This is a personal preference but I get frustrated with 'close', 'minimise', and 'maximise' suddenly appearing on the other side of the window and not looking like the rest of my native apps. I'm assuming this issue is more noticeable on certain Linux flavours and windows managers, and Mac OS.